# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: SeatShare ride-sharing platform API
servers:
  - url: /api
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/suspend:
    patch:
      operationId: suspendUser
      summary: Suspend or unsuspend a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /vehicles:
    get:
      operationId: listVehicles
      summary: List all vehicles
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleList"
    post:
      operationId: createVehicle
      summary: Register a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      summary: Get vehicle by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    patch:
      operationId: updateVehicle
      summary: Update vehicle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /trips:
    get:
      operationId: listTrips
      summary: List trips
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: driverProfileId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripList"
    post:
      operationId: createTrip
      summary: Driver publishes a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      summary: Get trip by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    patch:
      operationId: updateTrip
      summary: Update trip
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/cancel:
    patch:
      operationId: cancelTrip
      summary: Cancel a trip
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /ratings:
    get:
      operationId: listRatings
      summary: List ratings
      parameters:
        - in: query
          name: ratedId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingList"
    post:
      operationId: createRating
      summary: Rate a driver or passenger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /support-tickets:
    get:
      operationId: listSupportTickets
      summary: List support tickets
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketList"
    post:
      operationId: createSupportTicket
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support-tickets/{id}:
    get:
      operationId: getSupportTicket
      summary: Get support ticket by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
    patch:
      operationId: updateSupportTicket
      summary: Update support ticket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    UserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
    SuspendInput:
      type: object
      required: [suspend]
      properties:
        suspend:
          type: boolean
        reason:
          type: string
    DriverProfile:
      type: object
      required: [id, userId, isVerified, isOnline, rating, totalTrips, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        licenseNumber:
          type: ["string", "null"]
        rating:
          type: number
        totalTrips:
          type: integer
        isVerified:
          type: boolean
        isOnline:
          type: boolean
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        user:
          $ref: "#/components/schemas/User"
        vehicle:
          oneOf:
            - $ref: "#/components/schemas/Vehicle"
            - type: "null"
        createdAt:
          type: string
          format: date-time
...
    Vehicle:
      type: object
      required: [id, driverProfileId, make, model, year, color, licensePlate, capacity, createdAt]
      properties:
        id:
          type: integer
        driverProfileId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        capacity:
          type: integer
        createdAt:
          type: string
          format: date-time
    VehicleList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    VehicleInput:
      type: object
      required: [driverProfileId, make, model, year, color, licensePlate, capacity]
      properties:
        driverProfileId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        capacity:
          type: integer
    VehicleUpdate:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        capacity:
          type: integer
    Trip:
      type: object
      required: [id, driverProfileId, originAddress, destAddress, originLat, originLng, destLat, destLng, availableSeats, maxDeviationKm, status, farePerSeat, departureTime, createdAt]
      properties:
        id:
          type: integer
        driverProfileId:
          type: integer
        originAddress:
          type: string
        destAddress:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
        availableSeats:
          type: integer
        maxDeviationKm:
          type: number
        status:
          type: string
        farePerSeat:
          type: number
        departureTime:
          type: string
          format: date-time
        driverProfile:
          oneOf:
            - $ref: "#/components/schemas/DriverProfile"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    TripList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TripInput:
      type: object
      required: [driverProfileId, originAddress, destAddress, originLat, originLng, destLat, destLng, availableSeats, maxDeviationKm, farePerSeat, departureTime]
      properties:
        driverProfileId:
          type: integer
        originAddress:
          type: string
        destAddress:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
        availableSeats:
          type: integer
        maxDeviationKm:
          type: number
        farePerSeat:
          type: number
        departureTime:
          type: string
          format: date-time
    TripUpdate:
      type: object
      properties:
        status:
          type: string
        availableSeats:
          type: integer
    CancelInput:
      type: object
      properties:
        reason:
          type: string
...
    Rating:
      type: object
      required: [id, bookingId, raterId, ratedId, score, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        raterId:
          type: integer
        ratedId:
          type: integer
        score:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RatingList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RatingInput:
      type: object
      required: [bookingId, raterId, ratedId, score]
      properties:
        bookingId:
          type: integer
        raterId:
          type: integer
        ratedId:
          type: integer
        score:
          type: integer
        comment:
          type: string
    SupportTicket:
      type: object
      required: [id, userId, subject, message, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        resolution:
          type: ["string", "null"]
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    SupportTicketList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SupportTicketInput:
      type: object
      required: [userId, subject, message]
      properties:
        userId:
          type: integer
        subject:
          type: string
        message:
          type: string
    SupportTicketUpdate:
      type: object
      properties:
        status:
          type: string
        resolution:
          type: string
...