# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeatShare ride-sharing platform API
 * OpenAPI spec version: 1.0.0
 */

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  color?: string;
  licensePlate?: string;
  capacity?: number;
}